requirements.txt
streamlit==1.41.1
pymongo[srv]>=4.6.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
pyomo>=6.8.0
pandas>=2.2.0
plotly>=5.24.0
openpyxl>=3.1.0
pyscipopt>=6.0.0
